section 6 code

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,22 +1,70 @@
-import { v4 as uuidv4 } from 'uuid'
+import bcrypt from 'bcryptjs'
+import getUserId from '../utils/getUserId'
+import generateToken from '../utils/generateToken'
+import hashPassword from '../utils/hashPassword'
 
 
 const Mutation = {
   async createUser(parent, args, { prisma }, info) {
-      return await prisma.mutation.createUser({ data: args.data }, info)
+      const password = await hashPassword(args.data.password)
+      const user = await prisma.mutation.createUser({
+        data: {
+          ...args.data,
+          password
+        }
+      })
+
+      return {
+        user,
+        token: generateToken(user.id)
+      }
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteUser({where: {id: args.id}}, info)
+  async login(parent, args, {prisma}, info) {
+    const user = await prisma.query.user({
+      where: {
+        email: args.data.email
+      }
+    })
+
+    if (!user) {
+      throw new Error('Unable to login')
+    }
+
+    const isMatch = await bcrypt.compare(args.data.password, user.password)
+
+    if (!isMatch) {
+      throw new Error('Unable to login')
+    }
+
+    return {
+      user,
+      token: generateToken(user.id)
+    }
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    return prisma.mutation.deleteUser({
+      where: {id: userId}},
+      info)
+  },
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    if (typeof args.data.password === 'string') {
+      args.data.password = await hashPassword(args.data.password)
+    }
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info)
   },
-  async createPost(parent, args, { prisma }, info) {
+  async createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createPost({
       data: {
         title: args.data.title,
@@ -24,87 +72,56 @@ const Mutation = {
         published: args.data.published,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         }
       }
     }, info)
-      // check author exists
-      // const userExists = db.users.some((user) => user.id === args.data.author)
-
-      // if (!userExists) {
-      //     throw new Error('User not found.')
-      // }
-
-      // const post = {
-      //     id: uuidv4(),
-      //     ...args.data
-      // }
-
-      // db.posts.push(post)
-
-      // if (args.data.published) {
-      //   pubsub.publish("post", {
-      //     post: {
-      //       mutation: 'CREATED',
-      //       data: post
-      //     }
-      //   })
-      // }
-
-      // return post
   },
-  deletePost(parent, args, { prisma }, info) {
-      return prisma.mutation.deletePost({where: {id: args.id}}, info)
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to delete post')
+    }
+
+    return prisma.mutation.deletePost({where: {id: args.id}}, info)
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request}, info) {
+    const userId = getUserId(request)
 
-      // const post = db.posts.find((post) => post.id === args.id)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
 
-      // const { data } = args
-      // const originalPost = { ...post }
+    const isPublished =  await prisma.exists.Post({
+      id: args.id,
+      published: true
+    })
 
-      // if (!post) {
-      //     throw new Error('Post not found.')
-      // }
+    if (!postExists) {
+      throw new Error('Unable to update post')
+    }
 
-      // if (typeof data.title === 'string') {
-      //   post.title = data.title
-      // }
+    if (isPublished && !args.data.published) {
+      await prisma.mutation.deleteManyComments({
+        where: {
+          post: {
+            id: args.id
+          }
+        }
+      })
+    }
 
-      // if (typeof data.body === 'string') {
-      //   post.body = data.body
-      // }
-
-      // if (typeof data.published === 'boolean') {
-      //   post.published = data.published
-
-      //   if (originalPost.published && !post.published) {
-      //     pubsub.publish('post',  {
-      //       post: {
-      //         mutation: 'DELETED',
-      //         data: originalPost
-      //       }
-      //     })
-      //   } else if (!originalPost.published && post.published) {
-      //     pubsub.publish('post',  {
-      //       post: {
-      //         mutation: 'CREATED',
-      //         data: post
-      //       }
-      //     })
-      //   } else if (post.published) {
-      //     // updated
-      //     pubsub.publish('post',  {
-      //       post: {
-      //         mutation: 'UPDATED',
-      //         data: post
-      //       }
-      //     })
-      //   }
-      // }
-
-      // return post
       return prisma.mutation.updatePost({
         where: {
           id: args.id
@@ -112,13 +129,24 @@ const Mutation = {
         data: args.data
       }, info)
   },
-  createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = await prisma.exists.Post({
+      id: args.data.post,
+      published: true
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to find post')
+    }
+
       return prisma.mutation.createComment({
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           },
           post: {
@@ -129,14 +157,40 @@ const Mutation = {
         }
       }, info)
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to delete comment')
+    }
+
       return prisma.mutation.deleteComment({
         where: {
           id: args.id
         }
       }, info)
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to update comment')
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
